+++ changelog.md
@@ -18,6 +18,8 @@ FORMAT: 1A
 
 ## Unreleased v1.0.15 (Target March 31, 2018)
 ---------------------
+* FIX: Copy operation no longer requires header for Content-Type to be set
+* ENH: Listing group objects can now be done using user as group.
 * TODO: [Issue 1083](https://github.com/DecipherNow/object-drive-server/issues/1083) ADX Container Image
 * TODO: [Issue 947](https://github.com/DecipherNow/object-drive-server/issues/947) [Issue 1079](https://github.com/DecipherNow/object-drive-server/issues/1079) Migrate to Circle CI 2.0
 * TODO: [Issue 1081](https://github.com/DecipherNow/object-drive-server/issues/1081) Remove userHasGroup filter from list group objects